+++ src/pl-text.h
@@ -57,7 +57,7 @@ typedef struct
 int	PL_get_text(term_t l, PL_chars_t *text, int flags);
 int	PL_unify_text(term_t term, term_t tail, PL_chars_t *text, int type);
 int	PL_unify_text_range(term_t term, PL_chars_t *text,
-			    unsigned int from, unsigned int len, int type);
+			    size_t from, size_t len, int type);
 
 int	PL_promote_text(PL_chars_t *text);
 int	PL_demote_text(PL_chars_t *text);
